endian conversion and other changes

+++ src/seismic_data_provider.h
@@ -23,28 +23,28 @@ class seismic_data_provider : public obj_base {
 public:
 	seismic_data_provider() { }
 
-	virtual std::string											text_header() = 0;
-	virtual std::shared_ptr<seismic_abstract_header>			bin_header() = 0;
-	virtual std::shared_ptr<seismic_header_map>					header_map() = 0;
-	virtual void												set_header_map(std::shared_ptr<seismic_header_map> map) = 0;
+	virtual std::string                                         text_header() = 0;
+	virtual std::shared_ptr<seismic_abstract_header>            bin_header() = 0;
+	virtual std::shared_ptr<seismic_header_map>                 header_map() = 0;
+	virtual void                                                set_header_map(std::shared_ptr<seismic_header_map> map) = 0;
 
-	virtual std::shared_ptr<seismic_trace_header>				trace_header(int index) = 0;
-	virtual std::shared_ptr<seismic_trace>						get_trace(int index) = 0;
+	virtual std::shared_ptr<seismic_trace_header>               trace_header(int index) = 0;
+	virtual std::shared_ptr<seismic_trace>                      get_trace(int index) = 0;
 
-	virtual void												preprocessing() = 0;
+	virtual void                                                preprocessing() = 0;
 
-	virtual std::shared_ptr<seismic_geometry_info>				get_geometry() = 0;
+	virtual std::shared_ptr<seismic_geometry_info>              get_geometry() = 0;
 
-	virtual std::vector<std::shared_ptr<seismic_trace>>			get_traces(seismic_line_info line) = 0;
-	virtual std::vector<std::shared_ptr<seismic_trace>>			get_traces(int start, int end) = 0;
-	virtual std::vector<std::shared_ptr<seismic_trace_header>>	get_headers(seismic_line_info line) = 0;
-	virtual std::vector<std::shared_ptr<seismic_trace_header>>	get_headers(int start, int end) = 0;
+	virtual std::vector<std::shared_ptr<seismic_trace>>         get_traces(seismic_line_info line) = 0;
+	virtual std::vector<std::shared_ptr<seismic_trace>>         get_traces(int start, int end) = 0;
+	virtual std::vector<std::shared_ptr<seismic_trace_header>>  get_headers(seismic_line_info line) = 0;
+	virtual std::vector<std::shared_ptr<seismic_trace_header>>  get_headers(int start, int end) = 0;
 	// get_traces(points) (interpolation)
 
-	virtual void												close() = 0;
-	virtual int													traces_count() = 0;
-	virtual int													samples_count() = 0;
-	virtual float												sample_interval() = 0;
+	virtual void                                                close() = 0;
+	virtual int64_t                                             traces_count() = 0;
+	virtual int                                                 samples_count() = 0;
+	virtual double                                              sample_interval() = 0;
 };
 
 #ifdef PYTHON
